tweaks for android renaming
