+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.balancewallet",
+    package = "com.rainbow",
 )
 
 android_resource(
     name = "res",
-    package = "com.balancewallet",
+    package = "com.rainbow",
     res = "src/main/res",
 )
