+++ android/app/src/main/debug/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BalanceWallet Debug</string>
+    <string name="app_name">Rainbow Debug</string>
 </resources>
