+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BalanceWallet</string>
+    <string name="app_name">Rainbow</string>
 </resources>
