+++ android/app/src/main/staging/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BalanceWallet Staging</string>
+    <string name="app_name">Rainbow Staging</string>
 </resources>
