+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'BalanceWallet'
+rootProject.name = 'Rainbow'
 include ':react-native-device-info'
 project(':react-native-device-info').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-device-info/android')
 include ':react-native-screens'
